Move getStores & setStore to Rapidez facade
