+++ src/Commands/InteractsWithElasticsearchCommand.php
@@ -2,7 +2,6 @@
 
 namespace Rapidez\Core\Commands;
 
-use Carbon\Carbon;
 use Cviebrock\LaravelElasticsearch\Manager as Elasticsearch;
 use Illuminate\Console\Command;
 use Rapidez\Core\Models\Store;
@@ -51,23 +50,5 @@ abstract class InteractsWithElasticsearchCommand extends Command
     public function deleteIndex(string $index): void
     {
         $this->elasticsearch->indices()->delete(['index' => $index]);
-    }
-
-    public function getStores()
-    {
-        $storeModel = config('rapidez.models.store');
-        return $this->argument('store')
-            ? $storeModel::where('store_id', $this->argument('store'))->get()
-            : $storeModel::all();
-    }
-
-    public function setStore(Store $store): void
-    {
-        config()->set('rapidez.store', $store->store_id);
-        config()->set('rapidez.website', $store->website_id);
-        $code = config('rapidez.models.store')::getCachedWhere(function ($store) {
-            return $store['store_id'] == config('rapidez.store');
-        })['code'];
-        config()->set('rapidez.store_code', $code);
     }
 }
